+++ crypto/asn1/tasn_dec.c
@@ -126,7 +126,8 @@ unsigned long ASN1_tag2bit(int tag)
  */
 
 ASN1_VALUE *ASN1_item_d2i(ASN1_VALUE **pval,
-		const unsigned char **in, long len, const ASN1_ITEM *it)
+			  const unsigned char **in, size_t len,
+			  const ASN1_ITEM *it)
 	{
 	ASN1_TLC c;
 	ASN1_VALUE *ptmpval = NULL;
@@ -1243,7 +1244,7 @@ static int asn1_check_tlen(long *olen, int *otag, unsigned char *oclass,
 	{
 	int i;
 	int ptag, pclass;
-	long plen;
+	size_t plen;
 	const unsigned char *p, *q;
 	p = *in;
 	q = p;
